# doc-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/stat/{username}/{owner}/{repo}": {
      "get": {
        "summary": "Получение статистики репозитория",
        "description": "Возвращает статистику по конкретному репозиторию пользователя, \n    включая языки, стек технологий и метрики коммитов.\n    Если есть в базе данных статистика, то возвращает ее.\n    Если нет, то возвращает статистику по репозиторию с GitHub,\n    предварительно клонируя его, если параметр target = 'loclal',\n    или использует апи GitHub, если параметр target = 'github'.\n    Затем сохраняет статистику в базу данных.",
        "operationId": "get_repo_stat_stat__username___owner___repo__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          },
          {
            "name": "target",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "local",
              "title": "Target"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRepoStat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actualize_stat/{username}/{owner}/{repo}": {
      "get": {
        "summary": "Актуализация статистики репозитория",
        "description": "Обновляет информацию о статистике репозитория в базе данных.",
        "operationId": "get_actualize_stat_actualize_stat__username___owner___repo__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRepoStat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/global_stat/{username}": {
      "get": {
        "summary": "Получение глобальной статистики пользователя",
        "description": "Объединяет все статистики репозиториев пользователя, которые есть в базе данных и возвращает глобальную статистику.",
        "operationId": "get_global_stat_global_stat__username__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGlobalStat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user_info/{username}": {
      "get": {
        "summary": "Получение информации о пользователе",
        "description": "Возвращает информацию о профиле пользователя на GitHub.",
        "operationId": "get_user_info_user_info__username__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Регистрация пользователя в системе",
        "description": "Регистрирует нового пользователя с логином и паролем.",
        "operationId": "post_create_acc_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Авторизация пользователя",
        "description": "Авторизует пользователя в системе с логином и паролем.",
        "operationId": "post_login_acc_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/command": {
      "post": {
        "summary": "Добавление команды к аккаунту",
        "description": "Добавляет команду к учетной записи пользователя.",
        "operationId": "post_command_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_command_command_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/command/delete": {
      "post": {
        "summary": "Удаление команды из аккаунта",
        "description": "Удаляет команду из учетной записи пользователя.",
        "operationId": "post_command_del_command_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_command_del_command_delete_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "Поиск пользователей по компетенциям",
        "description": "Ищет пользователей на основе заданных компетенций и возвращает список профилей.",
        "operationId": "get_search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/activity/{username}/{owner}/{repo}": {
      "get": {
        "summary": "Получение числовых характеристик коммитов",
        "description": "Возвращает числовые характеристики изменений коммитов в указанном репозитории.",
        "operationId": "get_activity_activity__username___owner___repo__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/post_analysis/{username}/{owner}/{repo}": {
      "post": {
        "summary": "Анализ репозитория с использованием Llama",
        "description": "Добавляет в базу данных анализ репозитория, используя модель Llama.",
        "operationId": "post_analysis_post_analysis__username___owner___repo__post",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountInfo": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login",
            "description": "Логин"
          },
          "command_list": {
            "items": {
              "$ref": "#/components/schemas/CommandInfo"
            },
            "type": "array",
            "title": "Command List",
            "description": "Список команд"
          },
          "favorites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Favorites",
            "description": "Избранные разработчики"
          }
        },
        "type": "object",
        "required": [
          "login",
          "command_list",
          "favorites"
        ],
        "title": "AccountInfo"
      },
      "AccountRegister": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login",
            "description": "Логин"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Пароль"
          }
        },
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "title": "AccountRegister"
      },
...
      "Body_post_command_command_post": {
        "properties": {
          "cred": {
            "$ref": "#/components/schemas/AccountRegister"
          },
          "command": {
            "$ref": "#/components/schemas/CommandInfo"
          }
        },
        "type": "object",
        "required": [
          "cred",
          "command"
        ],
        "title": "Body_post_command_command_post"
      },
...
      "CommandInfo": {
        "properties": {
          "command_name": {
            "type": "string",
            "title": "Command Name",
            "description": "Название команды"
          },
          "participants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Participants",
            "description": "Список участников"
          }
        },
        "type": "object",
        "required": [
          "command_name",
          "participants"
        ],
        "title": "CommandInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserGlobalStat": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Имя пользователя на GitHub"
          },
          "contributed_repos": {
            "type": "integer",
            "title": "Contributed Repos",
            "description": "Число репозиториев, к которым внесены вклады"
          },
          "commits_total": {
            "type": "integer",
            "title": "Commits Total",
            "description": "Общее количество коммитов"
          },
          "commits_per_day": {
            "type": "number",
            "title": "Commits Per Day",
            "description": "Среднее коммитов в день"
          },
          "commits_per_week": {
            "type": "number",
            "title": "Commits Per Week",
            "description": "Среднее коммитов в неделю"
          },
          "commits_per_year": {
            "type": "number",
            "title": "Commits Per Year",
            "description": "Среднее коммитов в год"
          },
          "average_commit_size": {
            "type": "number",
            "title": "Average Commit Size",
            "description": "Средний размер коммита"
          },
          "competencies": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserCompetencyProfile"
              }
            ],
            "description": "Компетенции пользователя"
          },
          "using_github_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Using Github Features",
            "description": "Используемые функции GitHub, напр. 'pull_requests'"
          },
          "prep_repos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prep Repos",
            "description": "Список репозиториев со статистикой в БД"
          }
        },
        "type": "object",
        "required": [
          "username",
          "contributed_repos",
          "commits_total",
          "commits_per_day",
          "commits_per_week",
          "commits_per_year",
          "average_commit_size",
          "competencies",
          "using_github_features",
          "prep_repos"
        ],
        "title": "UserGlobalStat",
        "example": {
          "average_commit_size": 115,
          "commits_per_day": 2.5,
          "commits_per_week": 17.5,
          "commits_per_year": 910,
          "commits_total": 1000,
          "competencies": {
            "backend": [
              {
                "name": "Django",
                "score": 8
              }
            ],
            "devops": [
              {
                "name": "Docker",
                "score": 7
              }
            ],
            "resume": "Experienced in backend and DevOps"
          },
          "contributed_repos": 20,
          "prep_repos": [
            "john_doe/repo1",
            "mark/repo2"
          ],
          "username": "john_doe",
          "using_github_features": [
            "pull_requests",
            "wiki"
          ]
        }
      },
      "UserInfo": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Имя пользователя на GitHub"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Полное имя"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Электронная почта"
          },
          "team_projects": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Projects",
            "description": "Количество командных проектов"
          },
          "solo_projects": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solo Projects",
            "description": "Количество личных проектов"
          },
          "solo_gist": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solo Gist",
            "description": "Количество публичных гистов"
          },
          "account_age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Age",
            "description": "Возраст аккаунта в годах"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "description": "Ссылка на аватар"
          },
          "html_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Url",
            "description": "Ссылка на профиль"
          },
          "followers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followers",
            "description": "Количество подписчиков"
          },
          "following": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Following",
            "description": "Число отслеживаемых пользователей"
          },
          "repos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Repos",
            "description": "Список репозиториев в формате '{owner}/{repo}'"
          }
        },
        "type": "object",
        "required": [
          "username",
          "repos"
        ],
        "title": "UserInfo",
        "example": {
          "account_age": 3,
          "avatar_url": "https://example.com/avatar.jpg",
          "email": "john@example.com",
          "followers": 100,
          "following": 50,
          "html_url": "https://github.com/john_doe",
          "name": "John Doe",
          "repos": [
            "jack/repo1",
            "sparrow/repo2"
          ],
          "solo_gist": 3,
          "solo_projects": 10,
          "team_projects": 5,
          "username": "john_doe"
        }
      },
      "UserRepoStat": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Имя пользователя на GitHub"
          },
          "repo_name": {
            "type": "string",
            "title": "Repo Name",
            "description": "Репозиторий в формате '{owner}/{repo}'"
          },
          "repo_html_url": {
            "type": "string",
            "title": "Repo Html Url",
            "description": "URL репозитория"
          },
          "using_github_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Using Github Features",
            "description": "Используемые функции GitHub, напр. 'issues'"
          },
          "commits_total": {
            "type": "integer",
            "title": "Commits Total",
            "description": "Общее количество коммитов"
          },
          "commits_per_day": {
            "type": "number",
            "title": "Commits Per Day",
            "description": "Среднее коммитов в день"
          },
          "commits_per_week": {
            "type": "number",
            "title": "Commits Per Week",
            "description": "Среднее коммитов в неделю"
          },
          "commits_per_year": {
            "type": "number",
            "title": "Commits Per Year",
            "description": "Среднее коммитов в год"
          },
          "average_commit_size": {
            "type": "number",
            "title": "Average Commit Size",
            "description": "Средний размер коммита"
          },
          "competencies": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserCompetencyProfile"
              }
            ],
            "description": "Компетенции разработчика"
          }
        },
        "type": "object",
        "required": [
          "username",
          "repo_name",
          "repo_html_url",
          "using_github_features",
          "commits_total",
          "commits_per_day",
          "commits_per_week",
          "commits_per_year",
          "average_commit_size",
          "competencies"
        ],
        "title": "UserRepoStat",
        "example": {
          "average_commit_size": 120.5,
          "commits_per_day": 1.5,
          "commits_per_week": 10.5,
          "commits_per_year": 365,
          "commits_total": 150,
          "competencies": {
            "backend": [
              {
                "name": "FastAPI",
                "score": 4
              }
            ],
            "frontend": [
              {
                "name": "SVG",
                "score": 8
              }
            ],
            "resume": "Backend developer with expertise in Python"
          },
          "repo_html_url": "https://github.com/john_doe/sample_repo",
          "repo_name": "john_doe/sample_repo",
          "username": "john_doe",
          "using_github_features": [
            "actions",
            "issues"
          ]
        }
      },
...
    }
  }
}